Play whack-a-mole with errors on reviewAssessment page

+++ modules/life/life.php
@@ -1,11 +1,10 @@
 <?php
 function life_scoring($copy, $mysqli){
 
-    if (($copy['life_1'] ==1)||($copy['life_2'] ==1)||($copy['life_3'] ==1)){ //if there was a positive response to any of the yes/no questions
+    if ((array_key_exists('life_1', $copy) && $copy['life_1'] ==1)||($copy['life_2'] ==1)||($copy['life_3'] ==1)){ //if there was a positive response to any of the yes/no questions
             echo "<div>As documented in the Life Attitudes Schedule, note the following issues:</div>";
             echo "<span style=\"color:red\">";
-        if ($copy['life_1'] ==1){
-
+        if (array_key_exists('life_1', $copy) && $copy['life_1'] ==1){
             echo "<p>The client noted a desire to kill him or herself.</p>";
         }
         if ($copy['life_2'] ==1){
